+++ src/cli/status.cpp
@@ -15,6 +15,7 @@
 
 #include <uenv/env.h>
 #include <uenv/parse.h>
+#include <uenv/telemetry.h>
 #include <util/strings.h>
 
 #include "help.h"
@@ -57,96 +58,45 @@ int status([[maybe_unused]] const status_args& args,
         return args.error_if_unset ? 1 : 0;
     }
 
-    // assume that the environment variables have been set, because this is
-    // implied by the call to in_uenv_session passing
-    std::string mount_desc =
-        settings.calling_environment.get("UENV_MOUNT_LIST").value();
-    std::string view_literal =
-        settings.calling_environment.get("UENV_VIEW").value_or("");
+    auto telemetry = uenv::telemetry_from_env(settings.calling_environment);
 
-    const auto resolved = resolve_uenv_args(mount_desc, settings.config.repos);
-    if (!resolved) {
-        term::error("could not interpret UENV_MOUNT_LIST: {}",
-                    resolved.error());
-        return 1;
-    }
-
-    // the UENV_VIEW environment variable is a comma-separated list of the form
-    //   mount:uenv-name:view-name
-    // concretise_uenv requires a comma-separated list of the form
-    //   uenv-name:view-name
-    std::optional<std::string> view_desc;
-    if (auto views = parse_env_view_description(view_literal)) {
-        std::string view_string = "";
-        for (auto& v : *views) {
-            view_string += fmt::format("{}:{},", v.uenv, v.name);
-        }
-        view_desc = std::move(view_string);
-    } else {
-        spdlog::warn("unable to parse UENV_VIEW environment variable '{}'",
-                     view_desc);
-    }
-    spdlog::debug("derived view description from UENV_VIEW {}", view_desc);
-
-    const auto env = concretise_env(resolved.value(), view_desc, mount_desc,
-                                    settings.config.repos, false);
-
-    if (!env) {
-        term::error("could not interpret environment: {}", env.error());
+    if (!telemetry) {
+        term::error("unable to find uenv status");
         return 1;
     }
 
     if (args.format == full) {
-        for (auto& [name, E] : env->uenvs) {
-            term::msg("{}:{}", color::cyan(name), color::white(E.mount_path));
-            if (E.description) {
-                term::msg("  {}", *E.description);
-            }
-            if (!E.views.empty()) {
-                term::msg("  {}:", color::white("views"));
-                for (auto& [name, view] : E.views) {
-                    const bool loaded =
-                        std::ranges::find_if(
-                            env->views, [name, uenv = E.name](auto& p) {
-                                return p.name == name && p.uenv == uenv;
-                            }) != env->views.end();
-                    std::string status =
-                        loaded ? color::yellow(" (loaded)") : "";
-                    term::msg("    {}{}: {}", color::cyan(name), status,
-                              view.description);
-                }
-            }
+        for (auto& E : *telemetry) {
+            const std::string image = E.label ? E.label.value() : E.sqfs;
+            term::msg(
+                "uenv  {}\n  image  {}\n  mount  {}\n  views  [{}]",
+                color::yellow(E.name), color::cyan(image), color::cyan(E.mount),
+                fmt::join(E.views | std::views::transform([](const auto& v) {
+                              return color::cyan(v);
+                          }),
+                          ", "));
         }
     } else if (args.format == views) {
-        // print `<name1>:<view1>|..|<nameN>:<viewN>`
-        std::unordered_map<std::string, std::vector<std::string>> uenv_views;
-        // make sure there is an entry also for a mounted uenv without activated
-        // view
-        for (auto x : env->uenvs) {
-            uenv_views.try_emplace(x.first, std::vector<std::string>{});
-        }
-        for (auto x : env->views) {
-            uenv_views.try_emplace(x.uenv, std::vector<std::string>{});
-            uenv_views[x.uenv].push_back(x.name);
-        }
-        auto name_views =
-            uenv_views | std::views::transform([](const auto& pair) {
-                const auto& [name, views] = pair;
-                if (views.size() == 0) {
-                    return fmt::format("{}", name);
-                }
-                return fmt::format("{}[{}]", name, fmt::join(views, ","));
-            });
-        term::msg("{}", fmt::join(name_views, ","));
-        return 0;
-    } else if (args.format == name) {
-        // print `<name1>|..|<nameN>`
         term::msg(
             "{}",
-            fmt::join(env->uenvs | std::views::transform([](const auto& pair) {
-                          return fmt::format("{}", pair.first);
-                      }),
+            fmt::join(*telemetry | std::views::transform([](const auto& u) {
+                if (u.views.empty()) {
+                    return fmt::format("{}", color::yellow(u.name));
+                }
+                return fmt::format(
+                    "{}[{}]", color::yellow(u.name),
+                    fmt::join(u.views |
+                                  std::views::transform([](const auto& v) {
+                                      return color::cyan(v);
+                                  }),
+                              ","));
+            }),
                       ","));
+    } else if (args.format == name) {
+        term::msg("{}", fmt::join(*telemetry |
+                                      std::views::transform(
+                                          [](const auto& u) { return u.name; }),
+                                  ","));
     }
 
     return 0;