+++ src/slurm/environ.cpp
@@ -40,50 +40,6 @@ bool in_slurm_uenv_session(const envvars::state& env) {
     return uenv::in_uenv_session(env) && env.get("SLURM_UENV");
 }
 
-std::vector<telemetry_data>
-telemetry_from_env(const envvars::state& calling_environment) {
-    if (const auto var = calling_environment.get("UENV_TELEMETRY")) {
-        if (const auto result = parse_telemetry(var.value())) {
-            return result.value();
-        }
-        slurm_error("UENV_TELEMETRY is set but could not be parsed; "
-                    "falling back to UENV_MOUNT_LIST");
-    }
-
-    if (const auto mounts_var = calling_environment.get("UENV_MOUNT_LIST")) {
-        const auto views = parse_env_view_description(
-            calling_environment.get("UENV_VIEW").value_or(""));
-        if (const auto mounts = uenv::parse_mount_list(*mounts_var)) {
-            std::vector<telemetry_data> telemetry;
-            for (auto& m : *mounts) {
-                telemetry_data T{};
-                T.mount = m.mount_path;
-                T.sqfs = m.sqfs_path;
-                T.digest = std::nullopt;
-                T.label = std::nullopt;
-                if (views) {
-                    for (auto& v : *views) {
-                        if (v.mount == m.mount_path) {
-                            T.views.push_back(v.name);
-                        }
-                    }
-                }
-                // TODO: we could peek inside the squashfs
-                T.name = "unknown";
-                telemetry.push_back(std::move(T));
-            }
-            return telemetry;
-        } else {
-            slurm_error("UENV_MOUNT_LIST is malformed: %s",
-                        mounts.error().message().c_str());
-        }
-    }
-
-    slurm_error("unable to derive uenv telemetry from the environment");
-
-    return {};
-}
-
 // Update the calling environment to apply the environment variable updates.
 // note: making this into a nice clean function that returns state that can be
 // used by the calling slurm plugin was too hard - because slurm requires that
@@ -129,7 +85,7 @@ void patch_slurm_environment(const uenv::env& environment,
     std::string repo_v =
         full_env.get("UENV_REPO").value_or(args.repo_description.value_or(""));
     std::string uenv_v =
-        full_env.get("UENV").value_or(args.uenv_description.value_or(""));
+        full_env.get("UENV_LABEL").value_or(args.uenv_description.value_or(""));
     ::setenv("SLURM_UENV", uenv_v.c_str(), 1);
     ::setenv("SLURM_UENV_VIEW", view_v.c_str(), 1);
     ::setenv("SLURM_UENV_REPO", repo_v.c_str(), 1);