+++ changelog.md
@@ -1,7 +1,16 @@
 # Changelog
 
+## 10.0.2
+
+- #163 [improvement] add label/squashfs information to uenv status output
+
+## 10.0.1
+
+- #160 [fix] `uenv status` bug when no views are loaded, and only print active information about the loaded environment
+
 ## 10.0.0
 
+- #151 [fix] change error to warning when meta data is not attached to images in a registry
 - #150 [fix] fix missing headers in gcc 16
 - #149 [feature] generate RPMs in github actions
 - #148 [feature] support advanced slurm workflows